[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -8,8 +8,8 @@ int main()
     ios::sync_with_stdio(0);
     cin.tie(0);
 
-    int a, b;
-    int sum;
+    long a, b;
+    long sum;
 
     cin >> a >> b;
